requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-jose==3.3.0
stripe==7.11.0 
Flask-Migrate==4.1.0
